requirements.txt
jupyter
pandas
osmnx
geopandas
scikit-learn
pyyaml

